Create Know MAC Addresses
To modify the script to only show the MAC addresses and serial numbers of specific computers, you can add a conditional statement to filter the results based on the computer name or IP address.

In this script, the `$computers` variable defines an array of computer names or IP addresses to query. The script then loops through each computer and uses the `Get-WmiObject` cmdlet to retrieve the network adapters and serial number of each computer. The results arefiltered based on the computer name or IP address, and the MAC addresses and serial numbers are displayed for each computer.